Получение даты времени через метод класса ActiveBot. Использование типа данных short для класса ChanceWord. Чуть лучшая сортировка слушателей по пакетам.

+++ common/src/main/java/me/re4erka/botyara/api/bot/word/random/answer/ChanceWord.java
@@ -1,4 +1,4 @@
 package me.re4erka.botyara.api.bot.word.random.answer;
 
-public record ChanceWord(int index, String text, int lowerLimit, int upperLimit) {
+public record ChanceWord(short index, String text, short lowerLimit, short upperLimit) {
 }
